Move the uncategorized hooks into the common folder
