+++ packages/core/src/hooks/common/usePageScroll.ts
@@ -12,8 +12,8 @@ import {
   useContext,
 } from 'react';
 
-import { PageContext, PageContextInterface } from '../context/PageContext';
-import { ScrollPosition } from '../components/PageProvider';
+import { PageContext, PageContextInterface } from '../../context/PageContext';
+import { ScrollPosition } from '../../components/PageProvider';
 
 import { IntersectionInfo } from './useIntersectionObservable';
 import { useSubscription } from './useSubscription';
