+++ packages/core/src/index.ts
@@ -4,8 +4,8 @@ export * from './context/PageContext';
 export * from './utils/getIntersectionObserver';
 export * from './utils/getStickyPosition';
 export * from './hooks/page/useActiveSectionTracker';
-export * from './hooks/useIntersectionObservable';
-export * from './hooks/usePageScroll';
-export * from './hooks/useSectionPosition';
+export * from './hooks/common/useIntersectionObservable';
+export * from './hooks/common/usePageScroll';
+export * from './hooks/common/useSectionPosition';
 export * from './hooks/section/useSectionRatio';
-export * from './hooks/useSubscription';
+export * from './hooks/common/useSubscription';
